Fix for overrides wrapping from minimum to maximum rate
Overrides can wrap around to the maximum rate if multiple CMD_OVERRIDE_xx_MINUS commands are queued while
already at, or near to, the minimum rate. This can occur when using an encoder if moved too quickly.

This happens because the arguments to plan_feed_override() and spindle_set_override() are unsigned. If
(for instance) the override rate is 10, and commands from 11 negative encoder ticks are enqueued before
processing, then the requested override wraps round to a large positive number, and is capped at the
max override value.

+++ planner.c
@@ -565,7 +565,7 @@ void plan_cycle_reinitialize ()
 }
 
 // Set feed overrides
-void plan_feed_override (uint_fast8_t feed_override, uint_fast8_t rapid_override)
+void plan_feed_override (int_fast16_t feed_override, int_fast16_t rapid_override)
 {
     if(sys.override.control.feed_rate_disable)
         return;
